Revert back to 4.0 hostname/IP mapping
The RDO Manager instructions say to map the FQDN to 127.0.0.1.  If you do this,
then "openstack undercloud install" will fail because it cannot resolve the
short hostname.  This is because attempts to resolve the FQDN will return
"localhost", and result in nothing for a domain name.

As a result, we will stick with the 4.0 method of mapping the FQDN to the
external IP address.