Network isolation work in progress
This commit fixes a variety of issues with the network template files,
and adds a script to launch overcloud deployment.

This should be merged, but overcloud deployment with network isolation is
still a work in progress at this point.

There are also notably several hard coded parameters in deploy-overcloud.sh
that need to be editable.