Implemented R730xd flex bay drive support
This change implement flex bay drive support for the R730xds.  This is what
it does functionally:

If you're assigning the storage role to a node and the node is an R730xd
then it finds the first RAID 1 and assumes that this is the RAID made by the
DTK.  It checks to see if the size of the RAID is unique among all physical
disks on the node (minus the disk that comprise the RAID).  If the size is
unique, then it set the root_device hints on the node in ironic to be the
size of the RAID.

If the size of the RAID is NOT unique among the physical disks, then nothing
is done because ironic will not be able to uniquely identify the drive to
install on.  In this case, because ironic does not have any hints to tell it
what drive to install the OS on, it will install the OS on the first drive
of size greater than 4 gigs, which for us will always be physical drive sda.