New tool to make it easier to access overcloud nodes via SSH
update_ssh_config.py is a new Director tool. It uses the "nova list" command
to find the IP address assigned to each node, and creates a ~/.ssh/config
file that makes it easy to access each node. A simple name is used for each
node type ('X' is the node number):
  Node Name                Simple Name
  ------------------------ -----------
  overcloud-controller-X   cntlX
  overcloud-novacompute-X  novaX
  overcloud-cephstorage-X  storX

Also, it automatically logs in as the "heat-admin" user. So, instead of
doing this:

  nova list # To find the IP address assigned to overcloud-controller-0
  ssh heat-admin@192.168.120.YYY

You can do this:

  ssh cntl0