Incorporate review feedback on Ceph OSD configuration
Main change is an enhancement to the first-boot code that partitions the
disks so they can be used for Ceph. The new code supports storage nodes
with more than 26 block devices, and doesn't assume /dev/sdz is the last
one.

Also coverted ceph.yaml from DOS to UNIX line endings.