Configure proper pg_num for each Ceph pool
This is a workaround because upstream BZ 1283721 has been pushed to OSP-10.

Add new 'ceph_pools_pgs' to the ceph.yaml hieradata that provides a way to
specify the pg_num value for each pool. Update the post-deploy script with
code that reads the data, and increments each pool's pg_num and pgp_num
value until it matches.