Add ability to insert parameters into ceph.conf
This is a workaround until the TripleO heat templates support more Ceph
parameters. It implemenents a generic "ceph_tuning" hieradata entry in
ceph.yaml that defines a list of entries to be added to the [global] section
of the Cluster's ceph.conf file.

The only ceph_tuning parameter defines sets max_open_files = 131072. This
matches the corresponding JS-4.x setting. This is the only additional setting
warranted for JS-5. Other JS-4.x settings are obsolete or otherwise out of
favor in the community.