Fix for CES-4927: Boot disk issue on storage nodes
This patch changes assign_role so that for storage nodes, it always selects
the disk to install the OS on, and tells ironic which disk is the correct one.

Note that it is not necessary to do this for the other node types as they are
configured to have one huge volume by the DTK, and therefore there is only
one disk to install on.