Changes required to use Ceph radosgw for Swift object storage
Support for using the Ceph radosgw in OSP-8 is broken, and a number
of tweaks are required to fix things up so that it works.

Director:

o Create a framework for patching Director files such as Puppet code in
  /usr/share/openstack-tripleo-heat-templates. This is an interim workaround
  until they get their act together upstream. Adding a patched file to the
  overrides directory doesn't work because it always uses the Puppet code in
  /usr/share, and not the copy in pilot/templates/overcloud.
o Patch overcloud_controller_pacemaker.pp with the following changes:
  - Enable the RGW
  - Stop turning on the Swift proxy service (we're not using Swift)
  - Create the ceph.client.radosgw.gateway.keyring
o Enhance deploy-overcloud.py to update the Swift endpoint URLs with ones
  for the radosgw.
o Enhance credential_helper.py so it can also return the overcloud creds.

Templates:

o Disable Swift service by setting ControllerEnableSwiftStorage false in
  parameter_defaults, and swift::service::enabled false in hiera data. Both
  are necessary to disable the Swift services.
o Add hiera data to controllerExtraConfig to enable the radosgw with the
  civetweb front end.
o Add ceph_tuning entries to controller nodes to hook radosgw into keystone.

Deployment hooks:

o Add first-boot hook that patches the rgw.pp file in the overcloud. This
  is necessary to work around the broken "${name}" syntax that is supposed
  to expand to "radosgw.gateway".
o Patch the loadbalancer puppet file to use the radosgw's default port number
  for the HAProxy entry for the Swift service.
o Modify post-deploy code so that custom ceph_tuning can be applied to any
  node (not just storage nodes).
o Add post-deploy code so that the keystone URL and admin auth token is
  updated with the actual values.
o Add post-deploy code to restart the radosgw. This is necessary because the
  process is originally started too soon, before the radosgw keyring is
  present, and so the radosgw process picks up the newly patched keystone
  URL and auth token.