Fix for CES-5062: Auto SAH DHCP server interfering with Dir node
This fix enables and starts the DHCP server on the auto SAH node before
installing the SAH, and stops and disables the DHCP server after.
This prevents the DHCP server on the auto SAH from coming up if it
gets rebooted, which will reduce the chance that it will clash with
the director node.