Change ext network name from 'nova' to 'public'
OSP 8 creates default floating ip pool named public,
this used to be called nova.  There is a convention where
the external network name maps to the default connection
pool in nova.conf. Updaing the sanity test will keep this
in sync with OSP8 and make it easier to run tempest.