Fix for issue where adding a node fails
This patch fixes a potential issue where adding a node should fail.  The reason
for the failure is that we're picking out the line containing the heat stack
name from "heat stack-list" by assuming the deployment status contains a
certain string.  When rerunning a deployment, I don't believe the status will
contain that string, and while adding a node will succeed, the deployment will
appear to fail because we run update_ssh_config.py at the very end of
deployment, and this script gets the stack name and does stuff with it.  The
script won't be able to get the stack name, and so it will fail.

The fix is to change the way we get the stack name so that we use the heat
client API to get it directly instead of attempting to parse it from the
output of "heat stack-list".  Many thanks to Alan for this code!!!

I had to delete misc_helper.py and move the methods it contains into
credential_helper.py because of a circular import between the 2 files.

Also made a couple of very minor changes to split 1 line into 2 lines to make
some of the files pep8 compliant.

Also created an method to get the name of the undercloud rc file since we
have a method to get the name of the overcloud rc file.