Fix for CES-5919: Fix provisioning net VIP
In our solution, we state in several places that the control plane vip must be
in the dhcp_start, dhcp_end range since this is the info we were given by RH.
The vip cannot be the first IP in that range though because that IP is
allocated to the Neutron DHCP server that is running on the undercloud. The
storage management vip also cannot be the first IP in that range.

This patch fixes the range checking on the control plane and storage
management vips to prevent them from being the first IP in the range.
It also updates the sample .ini file with an appropriate comment.