Fix for CES-5945: Subscribe to and enable OSP repos outside deploy-director-vm.sh
This patch adds a script that the user will run on the director node
immediately after untaring the tar ball to subscribe the director node to the
OSP repos.

If any failure occurs, the script will display the error and exit immediately.

Note that we will have to update the repo names when they change.
Note that this script must either be run either as root or via sudo.

Usage:
    sudo ~/pilot/enable-repos.py