Fix Ceph RGW startup problem
On a newly created overcloud, the RGW would fail if it could not initialize
itself withing a configurable timeout period. The first time a Ceph cluster
is deployed, the RGW needs to create a couple of pools, and that would fail
if the storage nodes were busy creating the initial OSDs. The solution (also
implemented by other OpenStack distributions) is to extend the RGW
initialization timeout period from 5 to 20 minutes.