Fix networking glitch in SAH's kickstart file
The SAH's "anaconda_interface" is supposed to be disabled once the
installation is complete by adding "ONBOOT=no" to its ifcfg file.
However, due to a quirk in the sed script that updates the file, the
ONBOOT text would not be added if the file already contained an
ONBOOT directive. It would remove the old directive, but not append
the desired "ONBOOT=no". As a consequence, the interface would still
be active, and interfere with the SAH's br-extern interface.