Update method for setting Ceph PGs
OSP-10 now supports controller Ceph pool Placement Groups (PGs) on a per-pool
basis (the "CephPools" parameter). Changes to incorporate this include:
o Add a new "CephPools" element to the parameter_defaults
o Tweak the per-pool PG settings (includes adding a setting for a new
  OSP-10 pool called "backups")
o Remove the JetStream 6 code that was responsible for setting the PGs. That
  code is no longer necessary now that OSP-10 supports the feature.