Minor tweaks to deploy-overcloud.py
Fetch overcloud's SSH keys as soon as the deployment completes, and before
the finalize_overcloud() function is run. This ensures the keys are available
even if an error occurs when finalizing the overcloud.

Revise code that bypasses configuring the Swift protocol endpoint until such
time as the RGW is supported in OSP-10.