Deploy Storage Console VM instead of Ceph VM
The old Calamari UI for RHCS 1.3 has been replaced by the new Red Hat Storage
Console for RHCS 2.0. This commit essentially replaces the old "ceph" VM
that hosted the old Calamari UI with a new "rhscon" VM for the new Storage
Console.

A lot of the changes in this commit involves swapping the word "ceph" for
"rhscon" in the automation code.

This commit simply deploys the VM with the required software installed.
Additional work is required to configure the Storage Console, and hook it up
to the overcloud nodes.