Configure the Storage Console and overcloud Ceph nodes
Script the procedure for executing the commands that configures the Red Hat
Storage Console on the "rhscon" VM, and the Ceph Monitor and OSD nodes in
the overcloud. This prepares the overall system so that the user is able to
import the Ceph cluster into the Storage Console.

Portions of the code work around a few upstream bugs:
  o https://bugzilla.redhat.com/show_bug.cgi?id=1270860
  o https://bugzilla.redhat.com/show_bug.cgi?id=1414918
  o https://bugzilla.redhat.com/show_bug.cgi?id=1403576

The commit also include changes to our command_helper.py utility routines. I
refactored the code a little, and fixed an issue with the Scp file transfer
routines so that they work with key-based authentication sessions.