Refine script that configures the Storage Console
An upstream BZ (https://bugzilla.redhat.com/show_bug.cgi?id=1403576) is
currently preventing installation of the storage console agent on the
overcloud nodes unless the nodes are registered with CDN. The script that
configures the "rhscon" VM and the overcloud nodes used to detect the problem
and essentially abort.

This change introduces a small patch to the ceph-ansible software on the
rhscon VM that fixes the root problem. This allows the script to fully
configure the storage console, even when the overcloud nodes are not
subscribed to CDN.