Fix for RAID configuration hanging
This patch fixes the issue where when the disks are converted from RAID to
JBOD, the code hangs.

The response from the iDRAC sometimes comes back as OPTIONAL instead of YES.
In this case, the code assumed that reboot was not required, and would not
reboot the node.  The config job would then not run and our code would wait
forever for the job to complete.

In the end, if a disk is converted, then reboot is always required regardless
of what the iDRAC returns.