ces-7598: Automation: when exceptions occur, the cause may not be logged
https://cbjira.amer.dell.com/browse/CES-7598
When a command is executed remotely via paramiko and a pseudotty is used,
stderr is always returned as "", and the actual stderr is bundled in to stdout.
In most cases there is no need to use a pseudotty, and if you don't use one,
stderr and stdout are returned correctly.

Note that when a py script is executed remotely, anything logged with python
logging is dumped to stderr, not stdout.  Yes, even debug and info level
messages.

This patch makes the smallest changes possible to fix this problem for touch
free install so that errors are logged correctly.  At some later time, we
should try to remove as many cases of using the pseudotty as possible so
paramiko behaves as we expect it should by returning stderr and stdout
correctly.