Fix for CES-7619: director node appears to have lock on overcloud node blocking deployment
This patch increases the number of attempts to get the lock to 15.  This will
cause threads in ironic-conductor that are trying to get a lock on a node to
try to get the lock once a second for 15 seconds.  The end result is that the
IPA heartbeat will no longer timeout while trying to get a lock on a node when
RAID cleaning is in progress.

https://cbjira.amer.dell.com/browse/CES-7619