Select OS disk on all overcloud nodes
This patch changes assign_role so that it tells ironic what volume to install
the OS on for all overcloud nodes regardless of role.  The algorithm for
selecting the OS disk has been modified so that it looks for any RAID
volume that is not a RAID0, not just RAID1s.  This causes it to work for the
case where you have compute nodes with a RAID1+JBOD configuration, which is
what Hadoop on OpenStack has.