CES-7992: OSP 10 Puddle fails with waiting for DRAC to become ready
Three of the upstream patches to the iDRAC driver that we've submitted have
been incorporated into OSP10, which means that we can remove hand patching
the fixes for these bugs into the iDRAC driver.  The bios and wsman patches
are not longer needed as a result.

The lifecycle controller code has changed in the iDRAC driver so that the
existing patch will no longer apply.  I updated the patch so that it will
once again apply to the newer iDRAC driver lifecycle controller code.