Fix for ces-8606: assign_role on unconfigured 14g node fails raid config
This patch refactors introspect_nodes.py and adds introspect_node.py which is
to be used when adding a node to the cluster.  IB introspection has been
updated to check to see if local_gb was set by IB introspection, and launch
OOB introspection against all nodes where it has not been set.  This is to
handle RAID controllers that do not support passthrough where if no virtual
disks exist, then the controller does not expose any of the physical disks,
making it look like the node has no disks.