Fix for CES-8606: assign_role on unconfigured 14g node fails raid config
The root cause of this issue is that the PERC H740P RAID controller does
not support pass-through mode.  The end result of this is that only
virtual disks are made visible to the OS and physical disks are not
made visible to the OS.  As a result, in-band introspection will not
discover local_gb if there are no virtual disks.  This causes problems
further along in the flow.

This fix refactors introspec_nodes.py and adds an introspect_node.py
which will now be used in the add-a-node scenario.  By default,
introspect_node(s).py uses OOB introspection which side steps the
above issue.  If the user chooses to use IB introspection, then we
check to see if local_gb was discovered.  If it was not, then we run
OOB introspection to populate it.