Temporary fix to make use_custom_instack work
Currently use_custom_instack doesn't work because the instackenv.json is
copied to the director node as root.  config_idrac then tries to edit the
file and that fails because the director user does not own the file.

This is a temporary fix to change ownership of the file to the director user.