Add update script for 10.2
Did some refactoring of the 10.1 script including some
simplification and removal of workarounds that are no longer
needed.

Removed files not needed for update, which means everything
other than the update script itself.

Refactored archive name in build so it will named Update.tar
and extract to a folder called "update" as we no longer do any
upgrading this seemed to be the correct path.