Fix deployment failure due to invalid characters from iDRAC
This patch fixes an issue where on some stamps, when enumerating
DCIM_ControllerView, the iDRAC can return invalid unprintable
characters in the DriverVersion field.

This patch catches an XMLSyntaxError generated by etree.fromstring() during
enumeration and when it occurs, it strips out the unprintable characters and
reparses the XML in addition to logging the offending XML and the exception.