Fix setting boot_mode and spurious IB introspection. (#223)
This patch fixes the import_nodes.py script so that it allows the
boot_mode specified in the capabilities in the instack file to
be set on the nodes in ironic.

It also removes running in-band introspection against the
overcloud nodes in the prep script.  This makes deployment
run faster since OOB introspection is now used.