Add support for Glance Backend Option
Define the backend to use for Glance in the ini file.
For cinder
glance_backend=cinder
Set to rbd to use ceph as backend.
glance_backend=rbd
Set to file to use file-based storage for images.
glance_backend=file
For Swift
Set to swift to use swift to store images
glance_backend=swift
And deploy overcloud