Handle tempest config if public network not avail
If sanity test is not run prior to attempting to
configure tempest an excption would occur when the
tempest configuration script was called.  This patch
insures that the network the config script requires
is present, if it is not then the user is instructed
to run the sanity test to create the network and
to rerun the deployment with the --tempest_config_only
flag.

Also added --tempest_config_only flag to help user
recover from error where tempest.conf could not be
generated due to the sanity test failing or not
running at all, as executing it is currently an
option.

Fixes: CES-10320