Remove T10PI workaround
This patch removes the workaround for the T10PI issue since the
permanent fix has been released and verified in the HBA330 firmware
version 16.17.01.00.

Note that JetStream and Ironic users will need to upgrade the firmware
on their HBA330s to ensure that they do not encounter the T10PI issue
once this patch is merged.