Fix assign_role.py modification of node capabilities (#622)
Fix assign_role.py's modification of the Ironic bare metal node
'capabilities' property when it is already present. That can occur when
it is executed after out-of-band inspection or in-band introspection.
Both of those can set the 'boot_mode' capability on the node. Also,
'capabilities' can be set by a developer who executes JetPack scripts
manually, instead of through the automation.

When assign_role.py finds the node already has 'capabilities', it
modifies it by adding role and ensuring 'boot_mode' and 'boot_option'
are set to 'uefi' and 'local', respectively --
"boot_mode:uefi, boot_option:local". Otherwise, it adds the
'capabilities' property with the value set to role, "boot_mode:uefi",
and "boot_option:local".